+++ python_ipware/__init__.py
@@ -1,2 +1,6 @@
-from .python_ipware import IpWare  # noqa
-from .__version__ import __version__  # noqa
+from .__version__ import __version__
+from .legacy import LegacyIpWare
+from .modern import ModernIpWare
+from .router import IpWare
+
+__all__ = ["IpWare", "LegacyIpWare", "ModernIpWare", "__version__"]